+++ src/main/java/com/controle/estoque/hexagonal/dominio/portas/interfaces/ProdutoServicePort.java
@@ -1,9 +1,11 @@
 package com.controle.estoque.hexagonal.dominio.portas.interfaces;
 
+import com.controle.estoque.hexagonal.dominio.Produto;
 import com.controle.estoque.hexagonal.dominio.dtos.ProdutoDTO;
 import javassist.NotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface ProdutoServicePort {
 
@@ -11,5 +13,5 @@ public interface ProdutoServicePort {
 
     void salvarProduto(ProdutoDTO produtoDTO);
 
-    void atualizarEstoque(String id, Double qtdEstoque) throws NotFoundException;
+    void atualizarEstoque(String id, Integer qtdEstoque) throws NotFoundException;
 }
